+++ src/main/java/zero/eight/donut/controller/HomeReceiverController.java
@@ -8,30 +8,30 @@ import org.springframework.web.bind.annotation.RestController;
 import zero.eight.donut.common.response.SuccessResponse;
 import zero.eight.donut.config.annotation.LoginUser;
 import zero.eight.donut.domain.Receiver;
-import zero.eight.donut.dto.receiver.response.ReceiverGetBoxResponseDto;
-import zero.eight.donut.dto.receiver.response.ReceiverGetGiftResponseDto;
-import zero.eight.donut.dto.receiver.response.ReceiverHomeResponseDto;
+import zero.eight.donut.dto.home.receiver.ReceiverGetBoxResponseDto;
+import zero.eight.donut.dto.home.receiver.ReceiverGetGiftResponseDto;
+import zero.eight.donut.dto.home.receiver.ReceiverHomeResponseDto;
 import zero.eight.donut.exception.Success;
-import zero.eight.donut.service.ReceiverService;
+import zero.eight.donut.service.HomeReceiverService;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
-public class ReceiverController {
-    private  final ReceiverService receiverService;
+public class HomeReceiverController {
+    private  final HomeReceiverService homeReceiverService;
 
     @GetMapping("/home/receiver")
     public SuccessResponse<ReceiverHomeResponseDto> receiverHome(@LoginUser Receiver receiver){
-        return SuccessResponse.success(Success.HOME_RECEIVER_SUCCESS, receiverService.receiverHome(receiver));
+        return SuccessResponse.success(Success.HOME_RECEIVER_SUCCESS, homeReceiverService.receiverHome(receiver));
     }
 
     @GetMapping("/home/receiver/box/{boxId}")
     public SuccessResponse<ReceiverGetBoxResponseDto> receiverGetOneBox(@PathVariable("boxId") Long boxId){
-        return SuccessResponse.success(Success.HOME_RECEIVER_BOX_SUCCESS, receiverService.receiverGetOneBox(boxId));
+        return SuccessResponse.success(Success.HOME_RECEIVER_BOX_SUCCESS, homeReceiverService.receiverGetOneBox(boxId));
     }
 
     @GetMapping("/home/receiver/gift/{giftId}")
     public SuccessResponse<ReceiverGetGiftResponseDto> receiverGetOneGuft(@PathVariable("giftId") Long giftId){
-        return SuccessResponse.success(Success.HOME_RECEIVER_GIFT_SUCCESS, receiverService.receiverGetOneGift(giftId));
+        return SuccessResponse.success(Success.HOME_RECEIVER_GIFT_SUCCESS, homeReceiverService.receiverGetOneGift(giftId));
     }
 }
