+++ src/main/java/zero/eight/donut/service/HomeGiverService.java
@@ -3,7 +3,7 @@ package zero.eight.donut.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import zero.eight.donut.dto.giver.response.GiverHomeResponseDto;
+import zero.eight.donut.dto.home.giver.GiverHomeResponseDto;
 import zero.eight.donut.repository.DonationInfoRespository;
 import zero.eight.donut.repository.GiverRepository;
 import zero.eight.donut.repository.ReceiverRepository;
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
-public class GiverService {
+public class HomeGiverService {
     private final GiverRepository giverRepository;
     private final ReceiverRepository receiverRepository;
     private final DonationInfoRespository donationInfoRespository;
